+++ Sources/AvatarView.swift
@@ -74,7 +74,7 @@ open class AvatarView: UIView {
     }
 
     // MARK: - Initializers
-    override public init(frame: CGRect) {
+    public override init(frame: CGRect) {
         super.init(frame: frame)
         prepareView()
     }