+++ Sources/MediaMessageCell.swift
@@ -53,7 +53,7 @@ open class MediaMessageCell: MessageCollectionViewCell<UIImageView> {
         setupConstraints()
     }
 
-    override open func configure(with message: MessageType, at indexPath: IndexPath, and messagesCollectionView: MessagesCollectionView) {
+    open override func configure(with message: MessageType, at indexPath: IndexPath, and messagesCollectionView: MessagesCollectionView) {
         super.configure(with: message, at: indexPath, and: messagesCollectionView)
         switch message.data {
         case .photo(let image):