+++ Sources/MessageCollectionViewCell.swift
@@ -62,7 +62,7 @@ open class MessageCollectionViewCell<ContentView: UIView>: UICollectionViewCell 
 
     // MARK: - Initializer
 
-    override public init(frame: CGRect) {
+    public override init(frame: CGRect) {
         super.init(frame: frame)
         setupSubviews()
         setupGestureRecognizers()
@@ -85,7 +85,7 @@ open class MessageCollectionViewCell<ContentView: UIView>: UICollectionViewCell 
 
     }
 
-    override open func apply(_ layoutAttributes: UICollectionViewLayoutAttributes) {
+    open override func apply(_ layoutAttributes: UICollectionViewLayoutAttributes) {
         super.apply(layoutAttributes)
 
         guard let attributes = layoutAttributes as? MessagesCollectionViewLayoutAttributes else { return }
@@ -103,14 +103,14 @@ open class MessageCollectionViewCell<ContentView: UIView>: UICollectionViewCell 
 
     }
 
-    override open func prepareForReuse() {
+    open override func prepareForReuse() {
         cellTopLabel.text = nil
         cellTopLabel.attributedText = nil
         cellBottomLabel.text = nil
         cellBottomLabel.attributedText = nil
     }
 
-    public func configure(with message: MessageType, at indexPath: IndexPath, and messagesCollectionView: MessagesCollectionView) {
+    open func configure(with message: MessageType, at indexPath: IndexPath, and messagesCollectionView: MessagesCollectionView) {
 
         // Check if delegate has already been set to reduce number of assignments
         if delegate == nil, let cellDelegate = messagesCollectionView.messageCellDelegate {