+++ Sources/MessagesCollectionViewFlowLayout.swift
@@ -30,11 +30,7 @@ open class MessagesCollectionViewFlowLayout: UICollectionViewFlowLayout {
     // MARK: - Properties
 
     open var messageLabelFont: UIFont
-    open var messageLabelInsets: UIEdgeInsets
     open var messageToViewEdgePadding: CGFloat
-
-    open var cellTopLabelInsets: UIEdgeInsets
-    open var cellBottomLabelInsets: UIEdgeInsets
 
     open var avatarAlwaysLeading: Bool {
         willSet {
@@ -63,20 +59,16 @@ open class MessagesCollectionViewFlowLayout: UICollectionViewFlowLayout {
         return collectionView.frame.width - sectionInset.left - sectionInset.right
     }
 
-    override open class var layoutAttributesClass: AnyClass {
+    open override class var layoutAttributesClass: AnyClass {
         return MessagesCollectionViewLayoutAttributes.self
     }
 
     // MARK: - Initializers
 
-    override public init() {
+    public override init() {
 
         messageLabelFont = UIFont.preferredFont(forTextStyle: .body)
-        messageLabelInsets = UIEdgeInsets(top: 7, left: 14, bottom: 7, right: 14)
         messageToViewEdgePadding = 30.0
-
-        cellTopLabelInsets = .zero
-        cellBottomLabelInsets = .zero
 
         avatarAlwaysLeading = false
         avatarAlwaysTrailing = false
@@ -94,7 +86,7 @@ open class MessagesCollectionViewFlowLayout: UICollectionViewFlowLayout {
 
     // MARK: - Layout Attributes
 
-    override open func layoutAttributesForElements(in rect: CGRect) -> [UICollectionViewLayoutAttributes]? {
+    open override func layoutAttributesForElements(in rect: CGRect) -> [UICollectionViewLayoutAttributes]? {
 
         guard let attributesArray = super.layoutAttributesForElements(in: rect) as? [MessagesCollectionViewLayoutAttributes] else { return nil }
 
@@ -107,7 +99,7 @@ open class MessagesCollectionViewFlowLayout: UICollectionViewFlowLayout {
         return attributesArray
     }
 
-    override open func layoutAttributesForItem(at indexPath: IndexPath) -> UICollectionViewLayoutAttributes? {
+    open override func layoutAttributesForItem(at indexPath: IndexPath) -> UICollectionViewLayoutAttributes? {
 
         guard let attributes = super.layoutAttributesForItem(at: indexPath) as? MessagesCollectionViewLayoutAttributes else { return nil }
 
@@ -126,6 +118,9 @@ open class MessagesCollectionViewFlowLayout: UICollectionViewFlowLayout {
 
         let indexPath = attributes.indexPath
         let message = dataSource.messageForItem(at: indexPath, in: messagesCollectionView)
+        let messageInsets = messageLabelInsets(for: message, at: indexPath)
+        let topLabelInsets = cellTopLabelInsets(for: message, at: indexPath)
+        let bottomLabelInsets = cellBottomLabelInsets(for: message, at: indexPath)
 
         // First we set all the sizes so we can pass the attributes object around to calculate the origins
         attributes.messageContainerFrame = CGRect(origin: .zero, size: messageContainerSize(for: message, at: indexPath))
@@ -134,9 +129,9 @@ open class MessagesCollectionViewFlowLayout: UICollectionViewFlowLayout {
         attributes.avatarFrame = CGRect(origin: .zero, size: avatarSize(for: message, at: indexPath))
 
         attributes.messageLabelFont = messageLabelFont
-        attributes.messageLabelInsets = messageLabelInsets
-        attributes.cellTopLabelInsets = cellTopLabelInsets
-        attributes.cellBottomLabelInsets = cellBottomLabelInsets
+        attributes.messageLabelInsets = messageInsets
+        attributes.cellTopLabelInsets = topLabelInsets
+        attributes.cellBottomLabelInsets = bottomLabelInsets
         attributes.avatarHorizontalAlignment = avatarHorizontalAlignment(for: message)
 
         let layoutDelegate = messagesCollectionView.messagesLayoutDelegate
@@ -152,7 +147,7 @@ open class MessagesCollectionViewFlowLayout: UICollectionViewFlowLayout {
 
     // MARK: - Invalidation Context
 
-    override open func shouldInvalidateLayout(forBoundsChange newBounds: CGRect) -> Bool {
+    open override func shouldInvalidateLayout(forBoundsChange newBounds: CGRect) -> Bool {
 
         return collectionView?.bounds.width != newBounds.width || collectionView?.bounds.height != newBounds.height
 
@@ -260,14 +255,11 @@ extension MessagesCollectionViewFlowLayout {
 
     // MARK: - Message Container Calculations
 
-    /// The total value of the horizontal insets for the message label.
-    private var messageHorizontalInsets: CGFloat {
-        return messageLabelInsets.left + messageLabelInsets.right
-    }
-
-    /// The total value of the vertical insets for the message label.
-    private var messageVerticalInsets: CGFloat {
-        return messageLabelInsets.top + messageLabelInsets.bottom
+    /// The insets for the text of the MessageLabel
+    private func messageLabelInsets(for message: MessageType, at indexPath: IndexPath) -> UIEdgeInsets {
+        guard let messagesCollectionView = messagesCollectionView else { return .zero }
+        guard let layoutDelegate = messagesCollectionView.messagesLayoutDelegate else { return .zero }
+        return layoutDelegate.messageLabelInset(for: message, at: indexPath, in: messagesCollectionView)
     }
 
     /// The max width available for the message container.
@@ -278,6 +270,9 @@ extension MessagesCollectionViewFlowLayout {
         // Only apply padding if avatar width is greater than zero
         if avatarWidth > 0 { avatarWidth += avatarMessagePadding }
 
+        let messageInsets = messageLabelInsets(for: message, at: indexPath)
+        let messageHorizontalInsets = messageInsets.left + messageInsets.right
+
         let availableWidth = itemWidth - avatarWidth - messageToViewEdgePadding - messageHorizontalInsets
 
         return availableWidth
@@ -287,6 +282,9 @@ extension MessagesCollectionViewFlowLayout {
     fileprivate func messageContainerSize(for message: MessageType, at indexPath: IndexPath) -> CGSize {
 
         let maxWidth = messageContainerMaxWidth(for: message, at: indexPath)
+        let messageInsets = messageLabelInsets(for: message, at: indexPath)
+        let messageHorizontalInsets = messageInsets.left + messageInsets.right
+        let messageVerticalInsets = messageInsets.top + messageInsets.bottom
 
         var messageContainerSize: CGSize = .zero
 
@@ -340,14 +338,12 @@ extension MessagesCollectionViewFlowLayout {
 
     // MARK: - Cell Top Label Calculations
 
-    /// The total value of the horizontal insets for the cell top label.
-    private var topLabelHorizontalInsets: CGFloat {
-        return cellTopLabelInsets.left + cellTopLabelInsets.right
-    }
-
-    /// The total value of the vertical insets for the cell top label.
-    private var topLabelVerticalInsets: CGFloat {
-        return cellTopLabelInsets.top + cellTopLabelInsets.bottom
+    /// The insets for the cell's top label.
+    private func cellTopLabelInsets(for message: MessageType, at indexPath: IndexPath) -> UIEdgeInsets {
+        guard let messagesCollectionView = messagesCollectionView else { return .zero }
+        guard let layoutDelegate = messagesCollectionView.messagesLayoutDelegate else { return .zero }
+        let labelAlignment = layoutDelegate.cellTopLabelAlignment(for: message, at: indexPath, in: messagesCollectionView)
+        return labelAlignment.insets
     }
 
     /// The max width available for the cell top label.
@@ -359,6 +355,7 @@ extension MessagesCollectionViewFlowLayout {
         let labelHorizontal = layoutDelegate.cellTopLabelAlignment(for: message, at: indexPath, in: messagesCollectionView)
         let avatarVertical = layoutDelegate.avatarAlignment(for: message, at: indexPath, in: messagesCollectionView)
         let avatarHorizontal = avatarHorizontalAlignment(for: message)
+        let topLabelHorizontalInsets = labelHorizontal.insets.left + labelHorizontal.insets.right
 
         var avatarWidth = layoutDelegate.avatarSize(for: message, at: indexPath, in: messagesCollectionView).width
 
@@ -399,6 +396,10 @@ extension MessagesCollectionViewFlowLayout {
 
         let maxWidth = topLabelMaxWidth(for: message, at: indexPath)
         var size = labelSize(for: topLabelText, considering: maxWidth)
+
+        let topLabelInsets = cellTopLabelInsets(for: message, at: indexPath)
+        let topLabelHorizontalInsets = topLabelInsets.left + topLabelInsets.right
+        let topLabelVerticalInsets = topLabelInsets.top + topLabelInsets.bottom
 
         size.width += topLabelHorizontalInsets
         size.height += topLabelVerticalInsets
@@ -441,14 +442,12 @@ extension MessagesCollectionViewFlowLayout {
 
     // MARK: - Cell Bottom Label Calculations
 
-    /// The total value of the horizontal insets for the cell bottom label
-    private var bottomLabelHorizontalInsets: CGFloat {
-        return cellBottomLabelInsets.left + cellBottomLabelInsets.right
-    }
-
-    /// The total value of the vertical insets for the cell bottom label.
-    private var bottomLabelVerticalInsets: CGFloat {
-        return cellBottomLabelInsets.top + cellBottomLabelInsets.bottom
+    /// The insets for the cell's top label.
+    private func cellBottomLabelInsets(for message: MessageType, at indexPath: IndexPath) -> UIEdgeInsets {
+        guard let messagesCollectionView = messagesCollectionView else { return .zero }
+        guard let layoutDelegate = messagesCollectionView.messagesLayoutDelegate else { return .zero }
+        let labelAlignment = layoutDelegate.cellBottomLabelAlignment(for: message, at: indexPath, in: messagesCollectionView)
+        return labelAlignment.insets
     }
 
     /// The max with available for the cell bottom label.
@@ -460,6 +459,7 @@ extension MessagesCollectionViewFlowLayout {
         let labelHorizontal = layoutDelegate.cellBottomLabelAlignment(for: message, at: indexPath, in: messagesCollectionView)
         let avatarVertical = layoutDelegate.avatarAlignment(for: message, at: indexPath, in: messagesCollectionView)
         let avatarHorizontal = avatarHorizontalAlignment(for: message)
+        let bottomLabelHorizontalInsets = labelHorizontal.insets.left + labelHorizontal.insets.right
 
         var avatarWidth = layoutDelegate.avatarSize(for: message, at: indexPath, in: messagesCollectionView).width
 
@@ -501,6 +501,10 @@ extension MessagesCollectionViewFlowLayout {
 
         let maxWidth = bottomLabelMaxWidth(for: message, at: indexPath)
         var size = labelSize(for: bottomLabelText, considering: maxWidth)
+
+        let bottomLabelInsets = cellBottomLabelInsets(for: message, at: indexPath)
+        let bottomLabelHorizontalInsets = bottomLabelInsets.left + bottomLabelInsets.right
+        let bottomLabelVerticalInsets = bottomLabelInsets.top + bottomLabelInsets.bottom
 
         size.width += bottomLabelHorizontalInsets
         size.height += bottomLabelVerticalInsets